✨ Update music module

+++ src/main/java/fr/bakaaless/botzoe/bot/commands/music/SearchCommand.java
@@ -11,7 +11,9 @@ public class SearchCommand extends MusicCommand {
 
     @Override
     public void run(GuildMessageReceivedEvent event, String command, List<String> arguments) {
-        if (!super.canSendHere(event))
+        if (!super.canSendHere(event) || event.getMember() == null)
+            return;
+        if (!super.isInChannel(event.getMember(), true))
             return;
 
         event.getMessage().delete().queue();
